This is a Git commit message that summarizes the changes made to a project's configuration files:
```
export PATH="$PATH:/Users/ryanoatesmacminibase/.modular/bin"
```

The body of the commit message is empty, as the change is clear from the subject line and there is no additional relevant information to provide. The commit only adds the `export PATH="$PATH:/Users/ryanoatesmacminibase/.modular/bin"` line to the `.bashrc` file, which is a configuration file used by the Bash shell and sets the PATH environment variable.